# Allow qualified metadata references in the Update item operation