… src/Build/Evaluation/LazyItemEvaluator.LazyItemOperation.cs
… src/Build/Evaluation/LazyItemEvaluator.RemoveOperation.cs
… src/Build/Evaluation/LazyItemEvaluator.UpdateOperation.cs
… src/Build/Evaluation/LazyItemEvaluator.cs
… src/Shared/Traits.cs
… src/Shared/UnitTests/ObjectModelHelpers.cs