Resolves #[1655411 ](https://dev.azure.com/mseng/1ES/_workitems/edit/1655411), #[1649689](https://dev.azure.com/mseng/1ES/_workitems/edit/1649689)

This allows users to import and map metadata between two or more item types.

- Example:
```xml
   <I Update="@(Foo);*.cs;@(Bar)" M1="%(Foo.M1)" M2="%(Bar.X2)">
      <M3 Condition="'%(Foo.M3)'' == 'something' ">%(M1)</M3>
   </I>
```
- Each matching item (i.e. each item that needs uptading) now carries a context of captured item instances from all matching referenced item types in the `Update` expression.
- Unqualified metadata (`%(M)`) binds to the operation item type (i.e. type that gets updated). Qualified metadata (`%(Foo.M)`) binds inside the set of captured item types from the `Update` expression.
- If an item matches multiple times within and between multiple referenced items:
  - The last occurrence from each containing item type gets captured (so one captured item per item type)
  - This matches the behaviour of task item batching under targets. If we ever unify batching between evaluation and target items, there will hopefully be no semantic conflicts with this new form of Update.
- Where one can put `%()` references:
  - Metadata
  - Metadata conditions
- Breaking changes
```xml
    <ItemGroup>
        <OldUpdateBehaviour Include="a" M1="1" M2="x"/>
        <Subset Include="a" M1="newM1"/>
        <OldUpdateBehaviour Update="@(Subset)" M1="%(Subset.M1)" M2="%(Subset.M1)"/>
    </ItemGroup>
    <Target Name="Build">
        <Message Importance="High" Text="@(OldUpdateBehaviour) : M1 = [%(M1)] M2 = [%(M2)]"/>
    </Target>
```
Today this outputs: 
      `a : M1 = [] M2 = []`
After the change it will output: 
       `a : M1 = [newM1] M2 = [newM1]`

TODO:
- [x] investigate perf
- [x] finish some of the code todos