@jeffkl 2019-12-11
Thanks @cdmihai!  How does perf look so far?

/cc @cristinamanum