@cdmihai 2019-12-12
@jeffkl 
> How does perf look so far?

RPS [passed](https://devdiv.visualstudio.com/DevDiv/_git/VS/pullrequest/218598?_a=overview), so I didn't (observably) regress anything.

I also measured (via `msbuild.exe /profileEvaluation`) a simulated scenario (500 PackageVersion includes, 10 PackageReference includes (which match the PackageVersion items evenly)). I compared a master release build to this branch, on release. All the syntax from this PR is valid in master msbuild too, it just resolves to empty strings. So the difference is literally just the overhead of this change.
- for one update, item evaluation went up from 130ms to 162ms
- for 1000 updates (simulating 1000 projects) item evaluation went up from 14106ms to 28216ms

I'll work on some obvious optimizations. Beyond that, we need to define what's good enough.