@cdmihai 2019-12-17
I made some optimizations to reduce the overhead. 
- the 1000 updates scenario from above is now 300ms slower instead of twice as slow. At the existing 14 seconds, I think 300ms is acceptable
- I also ensured that total evaluation time for OrchardCore is unchanged. Prior to this optimization the new update made it 1 second slower (which makes me suspect RPS does not test any .net core scenarios)