@cdmihai 2019-12-17
@rainersigwald this probably warrants a mention in the [docs](https://docs.microsoft.com/en-us/visualstudio/msbuild/item-element-msbuild?view=vs-2019#attributes-and-elements), right?