@rainersigwald 2020-02-19
This definitely deserves a mention in the docs. Can you draft a PR for that too? I think it'll help contextualize the change.