@rainersigwald 2020-03-03
@cristinamanum fyi, this went in. It should be available in VS soonish.